requirements (1).txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.11.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.0
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
scikit-learn
matplotlib
python-dotenv